move graph making to rules, and integrate w/ validator
